refactor: 초기 레이아웃 세팅 부분 함수 분리
- 사실, 평큐때 들었던 대로 jsx와 비슷하게 보이게 머리를 굴려봤는데 지금은 도저히 생각이 안난다..
- 어차피 심화과제로 넘어갈 때 view영역은 react에게 위임할 것이니 이부분은 고민을 그만 멈추기로 했다.

- main.basic.js에 전역상태로 저장중인 UI 요소들도 굉장히 마음에 안들지만 위에 적은 이유와 마찬가지로 이부분도 고민을 멈추고, 일단 냅두기로 했다.
